Split apart route choice module
