+++ tests/aequilibrae/paths/test_route_choice.py
@@ -12,7 +12,8 @@ import pyarrow as pa
 from typing import List, Tuple
 
 from aequilibrae import Project
-from aequilibrae.paths.route_choice_set import RouteChoiceSet, GeneralisedCOODemand
+from aequilibrae.paths.cython.route_choice_set import RouteChoiceSet
+from aequilibrae.paths.cython.coo_demand import GeneralisedCOODemand
 from aequilibrae.paths.route_choice import RouteChoice
 from aequilibrae.matrix import AequilibraeMatrix, Sparse
